pyproject.toml
[tool.poetry]
name = "chat-connector"
version = "0.1.0"
description = ""
authors = ["Georgii Aleksahin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
faker = "^24.4.0"
mypy = "^1.9.0"
isort = "^5.13.2"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylint.design]
max-args = 10
max-locals = 15
max-attributes = 20
max-parents = 20
min-public-methods = 0
max-public-methods = 10
max-branches = 15
max-statements = 50
max-bool-expr = 10
max-returns = 10

[tool.pylint.format]
max-line-length = 100
max-module-lines = 500

[tool.pytest.ini_options]
addopts = "--disable-warnings"

[tool.pylint."messages control"]
confidence = [
    "HIGH",
    "CONTROL_FLOW",
    "INFERENCE",
    "INFERENCE_FAILURE",
    "UNDEFINED",
]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "unused-argument",
    "redefined-builtin",
    "import-outside-toplevel",
    "no-else-return",
    "unspecified-encoding",
    "duplicate-code",
    "super-with-arguments",
    "unsubscriptable-object",
    "abstract-method",
]

[tool.pylint.typecheck]
ignored-classes = [
    "optparse.Values",
    "thread._local",
    "_thread._local",
    "argparse.Namespace",
]
